+++ src/types.js
@@ -1,16 +1,27 @@
-const { array, boolean, defaulted, func, number, object, optional, record, string } = require('superstruct');
+const {
+  array,
+  assign,
+  boolean,
+  defaulted,
+  dynamic,
+  enums,
+  func,
+  number,
+  object,
+  optional,
+  record,
+  string
+} = require('superstruct');
 
-const OptionsStruct = defaulted(
+const OptionsStructBase = defaulted(
   object({
     enabled: optional(boolean()),
+    provider: optional(enums(['aws', 'elastic.co', 'vanilla'])),
     query: string(),
     selector: func(),
     toDocument: func(),
     endpoint: string(),
     index: string(),
-
-    accessKeyId: optional(string()),
-    secretAccessKey: optional(string()),
 
     mapping: record(
       string(),
@@ -25,8 +36,39 @@ const OptionsStruct = defaulted(
     plugins: array()
   }),
   {
-    enabled: false
+    enabled: false,
+    provider: 'aws'
   }
 );
+
+const OptionsStructAWS = assign(
+  OptionsStructBase,
+  object({
+    accessKeyId: string(),
+    secretAccessKey: string()
+  })
+);
+
+const OptionsStructElasticCo = assign(
+  OptionsStructBase,
+  object({
+    apiKey: string()
+  })
+);
+
+const DEFAULT_OPTIONS = OptionsStructAWS;
+
+const OptionsStruct = dynamic((options) => {
+  switch (options.provider) {
+    case 'aws':
+      return OptionsStructAWS;
+    case 'elastic.co':
+      return OptionsStructElasticCo;
+    case 'vanilla':
+      return OptionsStructBase;
+    default:
+      return DEFAULT_OPTIONS;
+  }
+});
 
 module.exports = { OptionsStruct };